requirements.txt
flask==3.0.2
requests==2.31.0
google-generativeai==0.4.1
pytest==8.0.2
pytest-mock==3.12.0
python-dotenv==1.0.1
